Fix memory leak in JSBundlerPlugin and remove a couple JSC::Strong (#22488)
### What does this PR do?

Since `JSBundlerPlugin` did not inherit from `JSDestructibleObject`, it
did not call the destructor. This means it never called the destructor
on `BundlerPlugin`, which means it leaked the WTF::Vector of RegExp and
strings.

This adds a small `WriteBarrierList` abstraction that is a
`WriteBarrier` guarded by the owning `JSCell`'s `cellLock()` that has a
`visitChildren` function. This also removes two usages of `JSC::Strong`
on the `Zig::GlboalObject` and replaces them with the
`WriteBarrierList`.

### How did you verify your code works?

Added a test. The test did not previously fail. But it's still good to
have a test that checks the onLoad callbacks are finalized.